requirements.txt
fastapi
uvicorn
openai
firebase-admin
requests
python-dotenv
pydantic
python-multipart
pillow
matplotlib
networkx
PyMuPDF
SpeechRecognition
pyaudio
elevenlabs
aiofiles
httpx
beautifulsoup4